+++ cpp/include/copc-lib/copc/file.hpp
@@ -20,7 +20,7 @@ class CopcFile : public las::LasFile
     CopcFile(const las::LasHeader &header, const las::CopcInfoVlr &copc_info, const CopcExtents &copc_extents,
              const las::WktVlr &wkt, const las::EbVlr &eb)
         : LasFile(header, eb), copc_info_(copc_info), copc_extents_(copc_extents), wkt_(wkt){};
-    CopcFile(const las::LasHeader &header, int spacing, const std::string &wkt, const las::EbVlr &eb)
+    CopcFile(const las::LasHeader &header, double spacing, const std::string &wkt, const las::EbVlr &eb)
         : LasFile(header, eb), copc_extents_(header.point_format_id, eb.items.size())
     {
         this->wkt_.wkt = wkt;