+++ test/points_test.cpp
@@ -119,7 +119,7 @@ TEST_CASE("Points tests", "[Point]")
 
     SECTION("Points Group Accessors")
     {
-        auto points = Points(3, copc::Vector3::DefaultScale(), copc::Vector3::DefaultOffset(), 4);
+        auto points = Points(7, copc::Vector3::DefaultScale(), copc::Vector3::DefaultOffset(), 4);
 
         // generate points
         int num_points = 2000;
@@ -129,6 +129,8 @@ TEST_CASE("Points tests", "[Point]")
             p->X(i);
             p->Y(i * 3);
             p->Z(i - 80);
+            p->Classification(i * 255 / num_points);
+            p->PointSourceID(i * 255 / num_points);
             points.AddPoint(p);
         }
 
@@ -138,42 +140,58 @@ TEST_CASE("Points tests", "[Point]")
         auto X = points.X();
         auto Y = points.Y();
         auto Z = points.Z();
+        auto classification = points.Classification();
+        auto point_source_id = points.PointSourceID();
         for (int i = 0; i < num_points; i++)
         {
             REQUIRE(X[i] == i);
             REQUIRE(Y[i] == i * 3);
             REQUIRE(Z[i] == i - 80);
+            REQUIRE(classification[i] == i * 255 / num_points);
+            REQUIRE(point_source_id[i] == i * 255 / num_points);
         }
 
         // generate vector of coordinates
         std::vector<double> Xn;
         std::vector<double> Yn;
         std::vector<double> Zn;
+        std::vector<uint8_t> classification_n;
+        std::vector<uint8_t> point_source_id_n;
 
         REQUIRE_THROWS(points.X(Xn));
         REQUIRE_THROWS(points.Y(Yn));
         REQUIRE_THROWS(points.Z(Zn));
+        REQUIRE_THROWS(points.Classification(classification_n));
+        REQUIRE_THROWS(points.PointSourceID(point_source_id_n));
 
         for (int i = 0; i < num_points - 1; i++)
         {
             Xn.push_back(i * 50 + 8);
             Yn.push_back(i + 800);
             Zn.push_back(i * 4);
+            classification_n.push_back(i * 255 / 2000);
+            point_source_id_n.push_back(i * 255 / 2000);
         }
 
         REQUIRE_THROWS(points.X(Xn));
         REQUIRE_THROWS(points.Y(Yn));
         REQUIRE_THROWS(points.Z(Zn));
+        REQUIRE_THROWS(points.Classification(classification_n));
+        REQUIRE_THROWS(points.PointSourceID(point_source_id_n));
 
         // add the last point
         Xn.push_back(1);
         Yn.push_back(2);
         Zn.push_back(3);
+        classification_n.push_back(255);
+        point_source_id_n.push_back(255);
 
         // test setters
         REQUIRE_NOTHROW(points.X(Xn));
         REQUIRE_NOTHROW(points.Y(Yn));
         REQUIRE_NOTHROW(points.Z(Zn));
+        REQUIRE_NOTHROW(points.Classification(classification_n));
+        REQUIRE_NOTHROW(points.PointSourceID(point_source_id_n));
 
         for (int i = 0; i < num_points - 1; i++)
         {
@@ -181,6 +199,8 @@ TEST_CASE("Points tests", "[Point]")
             REQUIRE(p->X() == i * 50 + 8);
             REQUIRE(p->Y() == i + 800);
             REQUIRE(p->Z() == i * 4);
+            REQUIRE(p->Classification() == i * 255 / 2000);
+            REQUIRE(p->PointSourceID() == i * 255 / 2000);
         }
 
         // test last point
@@ -188,6 +208,8 @@ TEST_CASE("Points tests", "[Point]")
         REQUIRE(last_point->X() == 1);
         REQUIRE(last_point->Y() == 2);
         REQUIRE(last_point->Z() == 3);
+        REQUIRE(last_point->Classification() == 255);
+        REQUIRE(last_point->PointSourceID() == 255);
     }
 
     SECTION("Points Indexers")
@@ -313,5 +335,29 @@ TEST_CASE("Points tests", "[Point]")
         points.AddPoint(p);
 
         REQUIRE(!points.Within(copc::Box(0, 0, 0, 5, 5, 5)));
+    }
+
+    SECTION("GetWithin")
+    {
+        auto points = Points(3, {1, 1, 1}, copc::Vector3::DefaultOffset());
+
+        std::random_device rd;
+        std::mt19937 gen(rd());
+        std::uniform_real_distribution<double> dist(0.0, 5.0);
+
+        // generate points
+        int num_points = 2000;
+        for (int i = 0; i < num_points; i++)
+        {
+            auto p = points.CreatePoint();
+            p->X(dist(gen));
+            p->Y(dist(gen));
+            p->Z(dist(gen));
+            points.AddPoint(p);
+        }
+        auto box = copc::Box(0, 0, 0, 2.5, 2.5, 2.5);
+        points = points.GetWithin(box);
+
+        REQUIRE(points.Within(box));
     }
 }