+++ CHANGELOG.md
@@ -7,6 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- **\[Python/C++\]** Add spatial function `Crosses` to `VoxelKey`
+- **\[Python/C++\]** Add resolution function `Resolution` and `GetResolutionAtDepth` to `VoxelKey`
+- **\[Python/C++\]** Add `Vector3` operators to `Vector3`
+- **\[Python/C++\]** Add resolution query functions `GetDepthAtResolution`, `GetNodesAtResolution` and `GetNodesWithinResolution` to `Reader`
+- **\[Python/C++\]** Add `BoundsTrimFileExample` and `ResolutionTrimFileExample` to example-writer files
+- **\[Python/C++\]** Add `Classification` and `PointSourceID` getter and setter to `Points`
+- **\[Python/C++\]** Add `GetChildren` to `VoxelKey`
+- **\[Python\]** Make `Node` and `VoxelKey` picklable
+
+### Changed
+
+- **\[Python/C++\]** Change order of arguments in `VoxelKey` spatial functions `Intersects`, `Contains`, and `Within`
+- **\[Python/C++\]** Add optional `resolution` argument to `Reader` spatial query functions `GetNodesWithinBox`, `GetNodesIntersectBox`, `GetPointsWithinBox`, and `GetAllPoints` . `resolution` can be used to limit the resolution during spatial queries
+- **\[Python/C++\]** Update `span` of `autzen-classified.copc.laz` test file from 0 to 128
+
+### Fixed
+
+- **\[Python\]** Fix typo in pickling of `Node`
+- **\[Python\]** Fix scale error and reformat `NewFileExample()` in example_writer.py
+
 ## [1.3.0] - 2021-10-04
 
 ### Added
@@ -57,6 +79,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Copc` class
 
 [1.1.0]: https://github.com/RockRobotic/copc-lib/compare/v1.0...v1.1.0
+[1.3.0]: https://github.com/RockRobotic/copc-lib/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/RockRobotic/copc-lib/compare/v1.1.0...v1.2.0
 [1.3.0]: https://github.com/RockRobotic/copc-lib/compare/v1.2.0...v1.3.0
-[unreleased]: https://github.com/RockRobotic/copc-lib/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/RockRobotic/copc-lib/compare/v1.3.0...HEAD